requirements.txt
fastapi
uvicorn
pillow
pdf2image
python-multipart
httpx
google-generativeai
python-dotenv
Pillow
markdown-pdf
markdown-it-py
PyMuPDF
marktex
pdfkit
markdown
sentence-transformers
langchain-huggingface
langchain
langchain-google-genai
accelerate
numpy
pytest
pytest-asyncio
